Prepare data source object for refactoring into API modules
