+++ src/modules/EcosystemDatabase.js
@@ -1,7 +1,13 @@
 import { SQLDataSource } from "datasource-sql";
 
 const MINUTE = 60;
+
 class EcosystemDatabase extends SQLDataSource {
+    // Extend the database object with a new API module
+    extend(apiModule, name) {
+        Object.setPrototypeOf(apiModule, this);
+        this[name] = apiModule;
+    }
 
     // ----------- User Auth Queries -----------
 
@@ -754,7 +760,6 @@ class EcosystemDatabase extends SQLDataSource {
             await trx.rollback()
         }
     }
-
 }
 
 export default EcosystemDatabase;
